requirements.txt
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.2
llvmlite==0.38.0
matplotlib==3.5.2
numba==0.55.1
numpy==1.21.6
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
